test: create folder for batching tests
